Add `insufficient_permissions_error`, `insufficient_permissions_error_request_access`, `data_loading_error` and `data_loading_error_retry` GA actions to AudienceSegmentationErrorWidget.
